# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Text editor MVP API
  version: 1.0.0
  description: Identity, Document and Processing contracts; ownership, folder and
    save semantics in docs.
servers:
  - url: http://localhost:8080
    description: Local same-origin reverse proxy
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/folders:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: listFolders
      summary: List immediate owned folder children
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderPage'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - &id002
          name: parent
          in: query
          required: false
          schema:
            type: string
            default: root
            maxLength: 36
          description: root or a UUID; null is not a query string literal.
        - &id003
          name: cursor
          in: query
          required: false
          schema:
            type: string
            maxLength: 2048
        - &id004
          name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
    post:
      tags:
        - Document
      x-service: document-service
      operationId: createFolder
      summary: Create an owned folder
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
  /api/v1/folders/{folderId}:
    get:
      tags:
        - Document
      x-service: document-service
      operationId: getFolder
      summary: Read own folder metadata
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: folderId
          in: path
          required: true
          schema: &id001
            type: string
            format: uuid
    patch:
      tags:
        - Document
      x-service: document-service
      operationId: patchFolder
      summary: Rename or move folder with revision check
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      description:
        At least one of name,parentId is required in addition to expectedMetadataRevision.
        Omitted parentId leaves unchanged; explicit null moves to root.
      parameters:
        - name: folderId
          in: path
          required: true
          schema: *id001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchFolderRequest'
    delete:
      tags:
        - Document
      x-service: document-service
      operationId: deleteFolder
      summary: Delete an empty owned folder
      responses:
        '204':
          description: Success with no body.
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
        '413':
          $ref: '#/components/responses/ErrorResponse'
        '422':
          $ref: '#/components/responses/ErrorResponse'
        '429':
          $ref: '#/components/responses/ErrorResponse'
        '503':
          $ref: '#/components/responses/ErrorResponse'
      parameters:
        - name: folderId
          in: path
          required: true
          schema: *id001
        - name: expectedMetadataRevision
          in: query
          required: true
          schema: &id005
            type: integer
            format: int64
            minimum: 1
            maximum: 9007199254740991
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          maxLength: 64
        message:
          type: string
          maxLength: 500
        traceId:
          type: string
          maxLength: 64
        details:
          type: object
          additionalProperties: true
      additionalProperties: false
      required:
        - code
        - message
        - traceId
...
    Folder:
      type: object
      properties:
        id: *id001
        parentId: &id007
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          maxLength: 120
        metadataRevision: *id005
        createdAt: &id006
          type: string
          format: date-time
        updatedAt: *id006
      additionalProperties: false
      required:
        - id
        - parentId
        - name
        - metadataRevision
        - createdAt
        - updatedAt
    CreateFolderRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        parentId: *id007
      additionalProperties: false
      required:
        - name
        - parentId
    PatchFolderRequest:
      type: object
      properties:
        expectedMetadataRevision: *id005
        name:
          type: string
          minLength: 1
          maxLength: 120
        parentId: *id007
      additionalProperties: false
      required:
        - expectedMetadataRevision
      minProperties: 2
...
    FolderPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Folder'
        nextCursor:
          type: string
          nullable: true
      additionalProperties: false
      required:
        - items
        - nextCursor
...
  responses:
    ErrorResponse:
      description: Error with stable code; semantics and status in docs/05_SAVE_AND_SHARE.md.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'